+++ partnerPercent.go
@@ -8,12 +8,12 @@ import (
 
 type PartnerPercentRule struct {
 	Id                 int     `gorm:"column:id"`
-	CarrierId          *int    `gorm:"column:carrier_id"`
+	CarrierId          *int64  `gorm:"column:carrier_id"`
 	Partner            *string `gorm:"column:partner"`
 	ConnectionGroup    *string `gorm:"column:connection_group"`
 	DateOfPurchaseFrom *string `gorm:"column:date_of_purchase_from"`
 	DateOfPurchaseTo   *string `gorm:"column:date_of_purchase_to"`
-	CarrierCountryId   *int    `gorm:"column:carrier_country_id"`
+	CarrierCountryId   *int64  `gorm:"column:carrier_country_id"`
 	FareType           *string `gorm:"column:fare_type"`
 	Result             float64 `gorm:"column:result"`
 	db                 *db.Database
